Fixed issue that caused incorrect sigma point generation. And adjusted initial Covariance matrix when the first sensor reading is from the laser.

+++ src/ukf_update.h
@@ -80,7 +80,7 @@ private:
                                     TAugSigmapointMatrix & augSigmaPoints);
 
   void PredictSigmaPoints(const TAugSigmapointMatrix& augSigmaPoints, 
-                          long delta_t, 
+                          double delta_t, 
                           TSigmapointsMatrix& sigmaPoints);
 
   void CalculateMeanAndCovariance(const TSigmapointsMatrix& predictedSigmaPoints,
